# docs/swagger.yaml
basePath: /
definitions:
  views.Swagger_Comment_Register_Post:
    properties:
      message:
        example: String
        type: string
      photo_id:
        example: 0
        type: integer
    type: object
  views.Swagger_Comment_Register_Put:
    properties:
      message:
        example: String
        type: string
    type: object
...
host: golang-final-project-production.up.railway.app
info:
  contact:
    email: jakaprima123@gmail.com
    name: Jaka Prima Maulana
  description: final project golang jaka
  termsOfService: http://swagger.io/terms/
  title: My Grams APP
  version: v2.0
paths:
  /comments:
    get:
      consumes:
      - application/json
      description: Get Comments
      parameters:
      - description: 'Token Barier example: ''Bearer 12355f32r'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      summary: Get Comments
      tags:
      - COMMENTS
    post:
      consumes:
      - application/json
      description: Post Comments
      parameters:
      - description: Post Comments
        in: body
        name: Comments
        required: true
        schema:
          $ref: '#/definitions/views.Swagger_Comment_Register_Post'
      - description: 'Token Barier example: ''Bearer 12355f32r'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      summary: Post Comments
      tags:
      - COMMENTS
  /comments/{commentId}:
    delete:
      consumes:
      - application/json
      description: Delete Comments
      parameters:
      - description: 'Token Barier example: ''Bearer 12355f32r'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID Comment
        in: path
        name: commentId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      summary: Delete Comments
      tags:
      - COMMENTS
    put:
      consumes:
      - application/json
      description: Update Comments
      parameters:
      - description: Update Comments
        in: body
        name: Comments
        required: true
        schema:
          $ref: '#/definitions/views.Swagger_Comment_Register_Put'
      - description: 'Token Barier example: ''Bearer 12355f32r'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID Comment
        in: path
        name: commentId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      summary: Update Comments
      tags:
      - COMMENTS
...
swagger: "2.0"